+++ docker-images/kafka/scripts/kafka_run.sh
@@ -38,7 +38,7 @@ if [ -e $KAFKA_HOME/rack/rack.id ]; then
 fi
 
 # Generate temporary keystore password
-export CERTS_STORE_PASSWORD=$(< /dev/urandom tr -dc _A-Z-a-z-0-9 | head -c${1:-32})
+export CERTS_STORE_PASSWORD=$(< /dev/urandom tr -dc _A-Z-a-z-0-9 | head -c32)
 
 mkdir -p /tmp/kafka